backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
httpx
psycopg2-binary

